Added Ability to open XML files succesfully
Instructions on testing the opening of XML files:

-place .XML in the Build folder(Example:
build-OSLD-Desktop_Qt_5_8_0_MSVC2015_64bit-Debug)[Note: Still needs to
be tested to see if it works in different folders]

-In DescriptionFileReader, change line 21 to the file name you want to
open. Do this before running the program. [Note: Will ad ability to do
this dynamically eventually]

Final Notes:
-The XML I am using to test can be found here:
http://www.xmlfiles.com/xml/xml_dtd.asp
(See "Well Formed XML Documents")

-The contents of the file are printed if the program is stopped. I still
need to look into why this is happening and if we can use this to our
advantage.